Add EVAL-WHEN to EXPORTs.
Ensure that exports have compile-time effect.
I'm not entirely sure this is necessary, because most people will use
ASDF in such a way that the protocol buffer compiled outputs will be
loaded before any downstream references.  But this is surely safer, and
it keeps the Allegro Common Lisp compiler from complaining.